feat: update relevant apis to scope based on role

+++ server/application/services/store.go
@@ -25,8 +25,11 @@ func (service *ProjectService) Create(params models.CreateProjectParams) (*model
 }
 
 // Details - Get project by id
-func (service *ProjectService) Details(id, orgID string) (*models.Project, error) {
-	return service.projectRepo.Details(context.Background(), id, orgID)
+func (service *ProjectService) Details(id, orgID, createdBy string, role models.Role) (*models.Project, error) {
+	if role == models.Admin {
+		return service.projectRepo.Details(context.Background(), id, orgID)
+	}
+	return service.projectRepo.DetailsByOrgAndCreator(context.Background(), id, orgID, createdBy)
 }
 
 func (service *ProjectService) GetProjectByID(id string) (*models.Project, error) {
@@ -39,8 +42,11 @@ func (service *ProjectService) Update(projectID string, params *models.UpdatePro
 }
 
 // Delete - Delete project
-func (service *ProjectService) Delete(id, orgID string) error {
-	return service.projectRepo.Delete(context.Background(), id, orgID)
+func (service *ProjectService) Delete(id, orgID, createdBy string, role models.Role) error {
+	if role == models.Admin {
+		return service.projectRepo.Delete(context.Background(), id, orgID)
+	}
+	return service.projectRepo.DeleteByOrgAndCreator(context.Background(), id, orgID, createdBy)
 }
 
 func (service *ProjectService) ProjectsBelongToOrg(projectIDs []string, orgID string) (bool, error) {
@@ -52,13 +58,17 @@ func (service *ProjectService) DatasetsBelongToOrg(datasetNames []string, orgID 
 }
 
 // List - Search projects
-func (service *ProjectService) List(query string, limit, page int, orgID string) (*models.PaginationView[*models.SearchProjectsResults], error) {
+func (service *ProjectService) List(query string, limit, page int, orgID, createdBy string, role models.Role) (*models.PaginationView[*models.SearchProjectsResults], error) {
 	pagination := models.NewPagination()
 	if limit != 0 {
 		pagination.Limit = limit
 	}
 	pagination.Offset = (page - 1) * limit
-	return service.projectRepo.SearchProject(context.Background(), query, pagination, orgID)
+
+	if role == models.Admin {
+		return service.projectRepo.SearchProject(context.Background(), query, pagination, orgID)
+	}
+	return service.projectRepo.SearchProjectByOrgAndCreator(context.Background(), query, pagination, orgID, createdBy)
 }
 
 func (service *ProjectService) ListAllProjects() ([]*models.Project, error) {
@@ -83,8 +93,11 @@ func (service *DatasetService) Create(params *models.CreateDatasetParams) (*mode
 	return service.datasetRepo.Create(context.Background(), params)
 }
 
-func (service *DatasetService) Details(id string, orgID string) (*models.Dataset, error) {
-	return service.datasetRepo.Details(context.Background(), id, orgID)
+func (service *DatasetService) Details(id, orgID, createdBy string, role models.Role) (*models.Dataset, error) {
+	if role == models.Admin {
+		return service.datasetRepo.Details(context.Background(), id, orgID)
+	}
+	return service.datasetRepo.DetailsByOrgAndCreator(context.Background(), id, orgID, createdBy)
 }
 
 func (service *DatasetService) GetByTableName(tableName string, orgID string) (*models.Dataset, error) {
@@ -100,8 +113,11 @@ func (service *DatasetService) List(projectID string, limit, page int) (*models.
 	return service.datasetRepo.List(context.Background(), projectID, pagination)
 }
 
-func (service *DatasetService) Delete(id string, orgID string) error {
-	return service.datasetRepo.Delete(context.Background(), id, orgID)
+func (service *DatasetService) Delete(id, orgID, createdBy string, role models.Role) error {
+	if role == models.Admin {
+		return service.datasetRepo.Delete(context.Background(), id, orgID)
+	}
+	return service.datasetRepo.DeleteByOrgAndCreator(context.Background(), id, orgID, createdBy)
 }
 
 func (service *DatasetService) Update(id string, params *models.UpdateDatasetParams) (*models.Dataset, error) {
